+++ src/Pim/Bundle/CatalogBundle/Model/ProductMedia.php
@@ -1,0 +1,14 @@
+<?php
+
+namespace Pim\Bundle\CatalogBundle\Model;
+
+/**
+ * Product media
+ *
+ * @author    Gildas Quemener <gildas@akeneo.com>
+ * @copyright 2013 Akeneo SAS (http://www.akeneo.com)
+ * @license   http://opensource.org/licenses/osl-3.0.php  Open Software License (OSL 3.0)
+ */
+class ProductMedia extends AbstractProductMedia
+{
+}