+++ src/Pim/Bundle/TranslationBundle/DependencyInjection/PimTranslationExtension.php
@@ -20,7 +20,6 @@ class PimTranslationExtension extends Extension
     public function load(array $configs, ContainerBuilder $container)
     {
         $loader = new Loader\YamlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
-        $loader->load('parameters.yml');
         $loader->load('form_types.yml');
         $loader->load('listeners.yml');
     }